pyproject.toml
# https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "courlan"
description = "Clean, filter and sample URLs to optimize data collection – includes spam, content type and language filters."
readme = "README.md"
license = { text = "Apache 2.0" }
dynamic = ["version"]
requires-python = ">=3.8"
authors = [
  {name = "Adrien Barbaresi", email = "barbaresi@bbaw.de"}
]
keywords=[
    "cleaner",
    "crawler",
    "uri",
    "url-parsing",
    "url-manipulation",
    "urls",
    "validation",
    "webcrawling",
]
classifiers = [
    # As from http://pypi.python.org/pypi?%3Aaction=list_classifiers
    'Development Status :: 5 - Production/Stable',
    #'Development Status :: 6 - Mature',
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Security",
    "Topic :: Text Processing :: Filters",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]
dependencies = [
    "babel >= 2.11.0",
    "tld >= 0.13",
    "urllib3 >= 1.26, < 3",
]

# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[tool.setuptools]
packages = ["courlan"]

# https://packaging.python.org/en/latest/guides/single-sourcing-package-version/
[tool.setuptools.dynamic]
version = {attr = "courlan.__version__"}

[project.scripts]
courlan = "courlan.cli:main"

[project.urls]
"Homepage" = "https://github.com/adbar/courlan"
"Blog" = "https://adrien.barbaresi.eu/blog/"  # /tag/courlan.html
"Tracker" = "https://github.com/adbar/courlan/issues"

# Development extras
[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "pytest",
    "pytest-cov",
]
